+++ e2e/testcases/v2-advanced-search/birth/1-search-birth-event-declaration-registration-details.spec.ts
@@ -1,9 +1,9 @@
 import { expect, test, type Page } from '@playwright/test'
-import { getToken, loginToV2 } from '../../helpers'
-import { createDeclaration } from '../v2-test-data/birth-declaration-with-father-brother'
-import { CREDENTIALS } from '../../constants'
+import { getToken, loginToV2 } from '../../../helpers'
+import { createDeclaration } from '../../v2-test-data/birth-declaration-with-father-brother'
+import { CREDENTIALS } from '../../../constants'
 import { faker } from '@faker-js/faker'
-import { type } from '../../v2-utils'
+import { type } from '../../../v2-utils'
 
 const todayDate = `${new Date().getDate() < 10 ? '0' : ''}${new Date().getDate().toString()}`
 const thisMonth = `${new Date().getMonth() < 9 ? '0' : ''}${(new Date().getMonth() + 1).toString()}`
