+++ e2e/testcases/v2-advanced-search/birth/2-search-birth-event-declaration-child-details.spec.ts
@@ -1,11 +1,11 @@
 import { expect, test, type Page } from '@playwright/test'
-import { getToken, loginToV2 } from '../../helpers'
-import { createDeclaration } from '../v2-test-data/birth-declaration-with-father-brother'
-import { CREDENTIALS } from '../../constants'
+import { getToken, loginToV2 } from '../../../helpers'
+import { createDeclaration } from '../../v2-test-data/birth-declaration-with-father-brother'
+import { CREDENTIALS } from '../../../constants'
 import { faker } from '@faker-js/faker'
-import { formatDateToLongString } from './utils'
-import { getMonthFormatted } from './helper'
-import { type } from '../../v2-utils'
+import { formatDateToLongString } from '../utils'
+import { getMonthFormatted } from '../helper'
+import { type } from '../../../v2-utils'
 
 test.describe
   .serial("Advanced Search - Birth Event Declaration - Child's details", () => {
