+++ e2e/testcases/v2-advanced-search/birth/4-search-birth-event-declaration-mother-details.spec.ts
@@ -1,20 +1,16 @@
 import { expect, test, type Page } from '@playwright/test'
-import { getToken, loginToV2 } from '../../helpers'
-import {
-  createDeclaration,
-  Declaration
-} from '../v2-test-data/birth-declaration-with-father-brother'
-import { CREDENTIALS } from '../../constants'
-import { formatDateToLongString } from './utils'
-import { faker } from '@faker-js/faker'
-import { getMonthFormatted } from './helper'
-import { type } from '../../v2-utils'
+import { getToken, loginToV2 } from '../../../helpers'
+import { createDeclaration } from '../../v2-test-data/birth-declaration-with-father-brother'
+import { CREDENTIALS } from '../../../constants'
+import { formatDateToLongString } from '../utils'
+import { getMonthFormatted } from '../helper'
+import { type } from '../../../v2-utils'
 
 test.describe
-  .serial("Advanced Search - Birth Event Declaration - Father's details", () => {
+  .serial("Advanced Search - Birth Event Declaration - Mother's details", () => {
   let page: Page
   let [yyyy, mm, dd] = ['', '', '']
-  let record: { eventId: string; declaration: Declaration }
+  let record: Awaited<ReturnType<typeof createDeclaration>>
 
   test.beforeAll(async ({ browser }) => {
     page = await browser.newPage()
@@ -22,8 +18,9 @@ test.describe
       CREDENTIALS.LOCAL_REGISTRAR.USERNAME,
       CREDENTIALS.LOCAL_REGISTRAR.PASSWORD
     )
-    record = await createDeclaration(token, {}, 'REGISTER', 'HEALTH_FACILITY')
-    ;[yyyy, mm, dd] = (record.declaration['father.dob'] ?? '').split('-')
+
+    record = await createDeclaration(token)
+    ;[yyyy, mm, dd] = record.declaration['mother.dob'].split('-')
   })
 
   test.afterAll(async () => {
@@ -37,28 +34,36 @@ test.describe
     await page.getByText('Birth').click()
   })
 
-  test.describe.serial("2.5 - Validate search by Father's details", () => {
+  test.describe.serial("2.5 - Validate search by Mother's details", () => {
     test('2.5.1 - Validate filling name and dob filters', async () => {
-      await page.getByText('Father details').click()
-
+      await page.getByText('Mother details').click()
       await type(
         page,
-        '#firstname',
-        record.declaration['father.name'].firstname ?? faker.person.firstName
+        '[data-testid="text__firstname"]',
+        record.declaration['mother.name'].firstname
       )
-      await type(page, '#surname', record.declaration['father.name'].surname)
-
-      await type(page, '[data-testid="father____dob-dd"]', dd)
-      await type(page, '[data-testid="father____dob-mm"]', mm)
-      await type(page, '[data-testid="father____dob-yyyy"]', yyyy)
+      await type(
+        page,
+        '[data-testid="text__surname"]',
+        record.declaration['mother.name'].surname
+      )
+      await type(page, '[data-testid="mother____dob-dd"]', dd)
+      await type(page, '[data-testid="mother____dob-mm"]', mm)
+      await type(page, '[data-testid="mother____dob-yyyy"]', yyyy)
     })
 
     test('2.5.2 - Validate search and show results', async () => {
       await page.click('#search')
       await expect(page).toHaveURL(/.*\/search-result/)
-      await expect(page.url()).toContain(`father.dob=${yyyy}-${mm}-${dd}`)
+      await expect(page.url()).toContain(`mother.dob=${yyyy}-${mm}-${dd}`)
       await expect(page.url()).toContain(
-        `father.name=${encodeURIComponent(JSON.stringify({ firstname: record.declaration['father.name'].firstname ?? faker.person.firstName, middlename: '', surname: record.declaration['father.name'].surname }))}`
+        `mother.name=${encodeURIComponent(
+          JSON.stringify({
+            firstname: record.declaration['mother.name'].firstname,
+            middlename: '',
+            surname: record.declaration['mother.name'].surname
+          })
+        )}`
       )
       await expect(page.getByText('Search results')).toBeVisible()
 
@@ -68,42 +73,43 @@ test.describe
       await expect(page.getByText('Event: V2 birth')).toBeVisible()
       await expect(
         page.getByText(
-          `Father's Date of birth: ${formatDateToLongString(record.declaration['father.dob'])}`
+          `Mother's Date of birth: ${formatDateToLongString(record.declaration['mother.dob'])}`
         )
       ).toBeVisible()
       await expect(
         page.getByText(
-          `Father's Name: ${record.declaration['father.name'].firstname ?? faker.person.firstName} ${record.declaration['father.name'].surname}`
+          `Mother's Name: ${record.declaration['mother.name'].firstname} ${record.declaration['mother.name'].surname}`
         )
       ).toBeVisible()
+
       await expect(page.getByRole('button', { name: 'Edit' })).toBeVisible()
     })
 
     test('2.5.3 - Validate clicking on the search edit button', async () => {
       await page.getByRole('button', { name: 'Edit' }).click()
       await expect(page).toHaveURL(/.*\/advanced-search/)
-      await expect(page.url()).toContain(`father.dob=${yyyy}-${mm}-${dd}`)
+      await expect(page.url()).toContain(`mother.dob=${yyyy}-${mm}-${dd}`)
       await expect(page.url()).toContain(
-        `father.name=${encodeURIComponent(JSON.stringify({ firstname: record.declaration['father.name'].firstname ?? faker.person.firstName, surname: record.declaration['father.name'].surname, middlename: '' }))}`
+        `mother.name=${encodeURIComponent(JSON.stringify({ firstname: record.declaration['mother.name'].firstname, surname: record.declaration['mother.name'].surname, middlename: '' }))}`
       )
       await expect(page.locator('#tab_v2\\.birth')).toHaveText('Birth')
-      await expect(page.getByTestId('father____dob-dd')).toHaveValue(dd)
-      await expect(page.getByTestId('father____dob-mm')).toHaveValue(mm)
-      await expect(page.getByTestId('father____dob-yyyy')).toHaveValue(yyyy)
+      await expect(page.getByTestId('mother____dob-dd')).toHaveValue(dd)
+      await expect(page.getByTestId('mother____dob-mm')).toHaveValue(mm)
+      await expect(page.getByTestId('mother____dob-yyyy')).toHaveValue(yyyy)
       await expect(page.locator('#firstname')).toHaveValue(
-        record.declaration['father.name'].firstname ?? faker.person.firstName
+        record.declaration['mother.name'].firstname
       )
       await expect(page.locator('#surname')).toHaveValue(
-        record.declaration['father.name'].surname
+        record.declaration['mother.name'].surname
       )
     })
 
-    test('2.5.4 - Validate father.dob range input', async () => {
-      const fatherDOBRangeButton = page.locator(
-        '#father____dob-date_range_button'
+    test('2.5.4 - Validate mother.dob range input', async () => {
+      const motherDOBRangeButton = page.locator(
+        '#mother____dob-date_range_button'
       )
-      if (await fatherDOBRangeButton.isVisible()) {
-        await page.locator('#father____dob-date_range_button').click()
+      if (await motherDOBRangeButton.isVisible()) {
+        await page.locator('#mother____dob-date_range_button').click()
         await expect(page.locator('#picker-modal')).toBeVisible()
 
         const currentMonth = new Date().getMonth() + 1
@@ -118,7 +124,7 @@ test.describe
         await expect(page.locator('#picker-modal')).toBeHidden()
 
         const checkbox = page.locator(
-          'input[type="checkbox"][name="father____dobdate_range_toggle"]'
+          'input[type="checkbox"][name="mother____dobdate_range_toggle"]'
         )
         await expect(checkbox).toBeVisible()
         await expect(checkbox).toBeChecked()
