+++ e2e/testcases/v2-advanced-search/birth/6-search-birth-event-declaration-informant-details.spec.ts
@@ -1,13 +1,13 @@
 import { expect, test, type Page } from '@playwright/test'
-import { getToken, loginToV2 } from '../../helpers'
-import { createDeclaration } from '../v2-test-data/birth-declaration-with-father-brother'
-import { CREDENTIALS } from '../../constants'
-import { formatDateToLongString } from './utils'
-import { getMonthFormatted } from './helper'
-import { type } from '../../v2-utils'
+import { getToken, loginToV2 } from '../../../helpers'
+import { createDeclaration } from '../../v2-test-data/birth-declaration-with-father-brother'
+import { CREDENTIALS } from '../../../constants'
+import { formatDateToLongString } from '../utils'
+import { getMonthFormatted } from '../helper'
+import { type } from '../../../v2-utils'
 
 test.describe
-  .serial("Advanced Search - Birth Event Declaration - Mother's details", () => {
+  .serial("Advanced Search - Birth Event Declaration - Informant's details", () => {
   let page: Page
   let [yyyy, mm, dd] = ['', '', '']
   let record: Awaited<ReturnType<typeof createDeclaration>>
@@ -18,9 +18,8 @@ test.describe
       CREDENTIALS.LOCAL_REGISTRAR.USERNAME,
       CREDENTIALS.LOCAL_REGISTRAR.PASSWORD
     )
-
-    record = await createDeclaration(token)
-    ;[yyyy, mm, dd] = record.declaration['mother.dob'].split('-')
+    record = await createDeclaration(token, {}, 'REGISTER', 'HEALTH_FACILITY')
+    ;[yyyy, mm, dd] = record.declaration['informant.dob'].split('-')
   })
 
   test.afterAll(async () => {
@@ -34,36 +33,28 @@ test.describe
     await page.getByText('Birth').click()
   })
 
-  test.describe.serial("2.5 - Validate search by Mother's details", () => {
+  test.describe.serial("2.5 - Validate search by Informant's details", () => {
     test('2.5.1 - Validate filling name and dob filters', async () => {
-      await page.getByText('Mother details').click()
+      await page.getByText('Informant details').click()
+
       await type(
         page,
-        '[data-testid="text__firstname"]',
-        record.declaration['mother.name'].firstname
+        '#firstname',
+        record.declaration['informant.name'].firstname
       )
-      await type(
-        page,
-        '[data-testid="text__surname"]',
-        record.declaration['mother.name'].surname
-      )
-      await type(page, '[data-testid="mother____dob-dd"]', dd)
-      await type(page, '[data-testid="mother____dob-mm"]', mm)
-      await type(page, '[data-testid="mother____dob-yyyy"]', yyyy)
+      await type(page, '#surname', record.declaration['informant.name'].surname)
+
+      await type(page, '[data-testid="informant____dob-dd"]', dd)
+      await type(page, '[data-testid="informant____dob-mm"]', mm)
+      await type(page, '[data-testid="informant____dob-yyyy"]', yyyy)
     })
 
     test('2.5.2 - Validate search and show results', async () => {
       await page.click('#search')
       await expect(page).toHaveURL(/.*\/search-result/)
-      await expect(page.url()).toContain(`mother.dob=${yyyy}-${mm}-${dd}`)
+      await expect(page.url()).toContain(`informant.dob=${yyyy}-${mm}-${dd}`)
       await expect(page.url()).toContain(
-        `mother.name=${encodeURIComponent(
-          JSON.stringify({
-            firstname: record.declaration['mother.name'].firstname,
-            middlename: '',
-            surname: record.declaration['mother.name'].surname
-          })
-        )}`
+        `informant.name=${encodeURIComponent(JSON.stringify({ firstname: record.declaration['informant.name'].firstname, middlename: '', surname: record.declaration['informant.name'].surname }))}`
       )
       await expect(page.getByText('Search results')).toBeVisible()
 
@@ -73,43 +64,42 @@ test.describe
       await expect(page.getByText('Event: V2 birth')).toBeVisible()
       await expect(
         page.getByText(
-          `Mother's Date of birth: ${formatDateToLongString(record.declaration['mother.dob'])}`
+          `Informant's Date of birth: ${formatDateToLongString(record.declaration['informant.dob'])}`
         )
       ).toBeVisible()
       await expect(
         page.getByText(
-          `Mother's Name: ${record.declaration['mother.name'].firstname} ${record.declaration['mother.name'].surname}`
+          `Informant's Name: ${record.declaration['informant.name'].firstname} ${record.declaration['informant.name'].surname}`
         )
       ).toBeVisible()
-
       await expect(page.getByRole('button', { name: 'Edit' })).toBeVisible()
     })
 
     test('2.5.3 - Validate clicking on the search edit button', async () => {
       await page.getByRole('button', { name: 'Edit' }).click()
       await expect(page).toHaveURL(/.*\/advanced-search/)
-      await expect(page.url()).toContain(`mother.dob=${yyyy}-${mm}-${dd}`)
+      await expect(page.url()).toContain(`informant.dob=${yyyy}-${mm}-${dd}`)
       await expect(page.url()).toContain(
-        `mother.name=${encodeURIComponent(JSON.stringify({ firstname: record.declaration['mother.name'].firstname, surname: record.declaration['mother.name'].surname, middlename: '' }))}`
+        `informant.name=${encodeURIComponent(JSON.stringify({ firstname: record.declaration['informant.name'].firstname, surname: record.declaration['informant.name'].surname, middlename: '' }))}`
       )
       await expect(page.locator('#tab_v2\\.birth')).toHaveText('Birth')
-      await expect(page.getByTestId('mother____dob-dd')).toHaveValue(dd)
-      await expect(page.getByTestId('mother____dob-mm')).toHaveValue(mm)
-      await expect(page.getByTestId('mother____dob-yyyy')).toHaveValue(yyyy)
+      await expect(page.getByTestId('informant____dob-dd')).toHaveValue(dd)
+      await expect(page.getByTestId('informant____dob-mm')).toHaveValue(mm)
+      await expect(page.getByTestId('informant____dob-yyyy')).toHaveValue(yyyy)
       await expect(page.locator('#firstname')).toHaveValue(
-        record.declaration['mother.name'].firstname
+        record.declaration['informant.name'].firstname
       )
       await expect(page.locator('#surname')).toHaveValue(
-        record.declaration['mother.name'].surname
+        record.declaration['informant.name'].surname
       )
     })
 
-    test('2.5.4 - Validate mother.dob range input', async () => {
-      const motherDOBRangeButton = page.locator(
-        '#mother____dob-date_range_button'
+    test('2.5.4 - Validate informant.dob range input', async () => {
+      const informantDOBRangeButton = page.locator(
+        '#informant____dob-date_range_button'
       )
-      if (await motherDOBRangeButton.isVisible()) {
-        await page.locator('#mother____dob-date_range_button').click()
+      if (await informantDOBRangeButton.isVisible()) {
+        await page.locator('#informant____dob-date_range_button').click()
         await expect(page.locator('#picker-modal')).toBeVisible()
 
         const currentMonth = new Date().getMonth() + 1
@@ -124,7 +114,7 @@ test.describe
         await expect(page.locator('#picker-modal')).toBeHidden()
 
         const checkbox = page.locator(
-          'input[type="checkbox"][name="mother____dobdate_range_toggle"]'
+          'input[type="checkbox"][name="informant____dobdate_range_toggle"]'
         )
         await expect(checkbox).toBeVisible()
         await expect(checkbox).toBeChecked()
