+++ e2e/testcases/v2-advanced-search/birth/7-search-birth-event-declaration-status.spec.ts
@@ -1,7 +1,7 @@
 import { expect, test, type Page } from '@playwright/test'
-import { joinValuesWith, loginToV2 } from '../../helpers'
+import { joinValuesWith, loginToV2 } from '../../../helpers'
 import { faker } from '@faker-js/faker'
-import { ensureOutboxIsEmpty, type } from '../../v2-utils'
+import { ensureOutboxIsEmpty, type } from '../../../v2-utils'
 
 test.describe
   .serial("Advanced Search - Birth Event Declaration - Informant's details", () => {
